Alow to update device.

+++ app/api_v1/devices.py
@@ -6,15 +6,18 @@ from Error import already_exists, validation, not_authorized
 
 from ..models.device import Device, db
 
-@api.route('/device/<int:id>', methods=['DELETE', 'GET', 'UPDATE'])
+@api.route('/device/<int:id>', methods=['DELETE', 'GET', 'POST'])
 def get_device(id):
 
     if not request.is_json:
         abort(404)
 
-    if request.method == 'DELETE':
+    device = Device.query.filter_by(id = id).first()
 
-        device = Device.query.filter_by(id = id).first()
+    if not device:
+        abort(400)
+
+    if request.method == 'DELETE':
 
         db.session.delete(device)
         db.session.commit()
@@ -22,10 +25,8 @@ def get_device(id):
         return jsonify({
             'status': 200
         }), 200
-        
-    if request.method == 'GET':
 
-        device = Device.query.filter_by(id = id).first()
+    if request.method == 'GET':
 
         return jsonify({
 
@@ -36,8 +37,30 @@ def get_device(id):
             'status': 201
         }), 201
 
-    if request.method == 'UPDATE':
-        print('UPDATE')
+    if request.method == 'POST':
+
+        name = request.json.get('name')
+        description = request.json.get('description')
+        pin = request.json.get('pin')
+
+        if name is None or description is None or pin is None:
+            return validation
+
+        device.name = name
+        device.description = description
+        device.pin = pin
+        device.state = False
+
+        db.session.commit()
+
+        return jsonify({
+
+            'name': device.name,
+            'description': device.description,
+            'pin': device.pin,
+            'state': device.state,
+            'status': 201
+        }), 201
 
 @api.route('/device', methods=['POST'])
 def create_device():
